requirements.txt
opencv-python
img2pdf
Pillow